Fetch anime by season and year
This commit introduces the ability to fetch anime based on a given season and year using the AllAnime API.

The `AnimeBySeasonYear` class handles the API request and parses the JSON response to create a list of `AnimeItem` objects.

In `MainActivity`, a coroutine is launched on the IO dispatcher to perform the network request. The results are then displayed in a RecyclerView on the main thread.